+++ src/main/java/eu/europa/ec/dgc/gateway/DgcGatewayApplication.java
@@ -18,9 +18,9 @@
  * ---license-end
  */
 
-package eu.dgc.gateway;
+package eu.europa.ec.dgc.gateway;
 
-import eu.dgc.gateway.config.DgcConfigProperties;
+import eu.europa.ec.dgc.gateway.config.DgcConfigProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
