+++ src/main/java/eu/europa/ec/dgc/gateway/config/SchedulerConfig.java
@@ -18,33 +18,16 @@
  * ---license-end
  */
 
-package eu.dgc.gateway.config;
+package eu.europa.ec.dgc.gateway.config;
 
-import static net.javacrumbs.shedlock.provider.jdbctemplate.JdbcTemplateLockProvider.Configuration.builder;
-
-import javax.sql.DataSource;
-import net.javacrumbs.shedlock.core.LockProvider;
-import net.javacrumbs.shedlock.provider.jdbctemplate.JdbcTemplateLockProvider;
-import org.springframework.context.annotation.Bean;
+import net.javacrumbs.shedlock.spring.annotation.EnableSchedulerLock;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.context.annotation.Profile;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
-public class ShedLockConfig {
-
-  /**
-   * Creates a LockProvider for ShedLock.
-   *
-   * @param dataSource JPA datasource
-   * @return LockProvider
-   */
-  @Bean
-  public LockProvider lockProvider(DataSource dataSource) {
-    return new JdbcTemplateLockProvider(builder()
-      .withTableName("shedlock")
-      .withJdbcTemplate(new JdbcTemplate(dataSource))
-      .usingDbTime()
-      .build()
-    );
-  }
+@Profile("!test")
+@EnableScheduling
+@EnableSchedulerLock(defaultLockAtMostFor = "PT30S")
+public class SchedulerConfig {
 }
